authentication_routes

Related work items: #25

+++ client/src/components/WorkoutDetails.js
@@ -1,12 +1,20 @@
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext';
 //date-fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
+import { useAuthContext } from '../hooks/useAuthContext';
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
   const handleClick = async () => {
+    if (!user) {
+      return;
+    }
     const response = await fetch('/api/workouts/' + workout._id, {
       method: 'DELETE',
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
     const json = await response.json();
     if (response.ok) {
